Minor style improvements
