+++ tests/schemas/unauthorized-schema.ts
@@ -1,0 +1,8 @@
+export const unauthorizedSchema = {
+  type: "object",
+  required: ["detail"],
+  additionalProperties: false,
+  properties: {
+    detail: { type: "string" },
+  },
+};
