--- tests/schemas/unauthorized.schema.ts
@@ -1,8 +1,0 @@
-export const unauthorizedSchema = {
-    type: "object",
-    required: ["detail"],
-    additionalProperties: false,
-    properties: {
-      detail: { type: "string" }
-    }
-  };
